refactor: move options and parsing outside main
