+++ src/tools/androiddeployqt/CMakeLists.txt
@@ -13,6 +13,9 @@ qt_internal_add_tool(${target_name}
     INSTALL_VERSIONED_LINK
     SOURCES
         main.cpp
+        options.h
+        OptionsParser.h
+        OptionsParser.cpp
     DEFINES
         QT_NO_CAST_FROM_ASCII
         QT_NO_CAST_TO_ASCII
